Remove dependency on utils package for kafka example
utils package is not currently open sourced and the simplest way of setting
up kafka variables is to hardcode them into application. If a developer
wants to use environment variables this can be done several different
ways but should not effect an example app of how to use this go-ns kafka library.